+++ R/add_extra_portfolio.R
@@ -112,14 +112,15 @@ add_extra_portfolio <- function(x) {
           ## compile results
           if (!is.null(sol$pool)) {
             sol <- append(
-              list(sol[-5]),
+              list(sol[-which(names(sol) == "pool")]),
               lapply(
                 sol$pool,
                 function(z) list(
                   x = z$xn,
-                  objective = z$objval,
+                  objective = z$objective,
                   status = z$status,
-                  runtime = sol$runtime
+                  runtime = sol$runtime,
+                  gap = z$gap
                 )
               )
             )