+++ R/add_linear_constraints.R
@@ -297,7 +297,7 @@ methods::setMethod("add_linear_constraints",
       assertthat::noNA(d),
       msg = paste(
         "{.arg data} must not refer to columns",
-        "of the planning unit data for {.arg} with missing",
+        "of the planning unit data with missing",
         "({.val {NA}) values."
       )
     )