+++ R/add_rsymphony_solver.R
@@ -206,7 +206,8 @@ add_rsymphony_solver <- function(x, gap = 0.1,
             x = x$solution,
             objective = x$objval,
             status = as.character(x$status),
-            runtime = rt[[3]]
+            runtime = rt[[3]],
+            gap = NA_real_
           )
         }
       )