+++ tests/testthat/test_add_gurobi_solver.R
@@ -18,8 +18,8 @@ test_that("binary decisions", {
     add_binary_decisions() %>%
     add_gurobi_solver(time_limit = 5, numeric_focus = TRUE, verbose = FALSE)
   # solve problems
-  s1 <- solve_fixed_seed(p1)
-  s2 <- solve_fixed_seed(p2)
+  s1 <- solve(p1)
+  s2 <- solve(p2)
   # tests
   expect_inherits(s1, "SpatRaster")
   expect_equal(terra::nlyr(s1), 1)
@@ -42,7 +42,7 @@ test_that("proportion decisions", {
     add_proportion_decisions() %>%
     add_gurobi_solver(gap = 0, verbose = FALSE)
   # solve problem
-  s <- solve_fixed_seed(p)
+  s <- solve(p)
   # tests
   expect_inherits(s, "SpatRaster")
   expect_equal(terra::nlyr(s), 1)
@@ -68,7 +68,7 @@ test_that("proportion decisions (floating point)", {
     add_proportion_decisions() %>%
     add_gurobi_solver(gap = 0, verbose = FALSE)
   # solve problem
-  s <- solve_fixed_seed(p)
+  s <- solve(p)
   # tests
   expect_inherits(s, "sf")
   expect_true("solution_1" %in% names(s))
@@ -120,7 +120,7 @@ test_that("mix of binary and continuous variables", {
     add_binary_decisions() %>%
     add_gurobi_solver(verbose = FALSE)
   # solve problem
-  s <- solve_fixed_seed(p)
+  s <- solve(p)
   # tests
   expect_inherits(s, "SpatRaster")
   expect_equal(terra::nlyr(s), 1)
@@ -142,7 +142,7 @@ test_that("first_feasible", {
     add_binary_decisions() %>%
     add_gurobi_solver(first_feasible = TRUE, verbose = FALSE)
   # solve problem
-  s1 <- solve_fixed_seed(p1)
+  s1 <- solve(p1)
   # check that solution has correct properties
   expect_inherits(s1, "SpatRaster")
   expect_equal(terra::nlyr(s1), 1)
@@ -171,8 +171,8 @@ test_that("correct solution (simple)", {
     add_locked_out_constraints(locked_out) %>%
     add_gurobi_solver(gap = 0, verbose = FALSE)
   # solve problems
-  s1 <- solve_fixed_seed(p)
-  s2 <- solve_fixed_seed(p)
+  s1 <- solve(p)
+  s2 <- solve(p)
   # tests
   expect_equal(c(terra::values(s1)), c(0, 1, 1, NA))
   expect_equal(terra::values(s1), terra::values(s2))
@@ -201,8 +201,8 @@ test_that("correct solution (complex)", {
         target = c(5, 10, 20))) %>%
     add_gurobi_solver(gap = 0, verbose = FALSE)
   # solve problems
-  s1 <- solve_fixed_seed(p)
-  s2 <- solve_fixed_seed(p)
+  s1 <- solve(p)
+  s2 <- solve(p)
   # tests
   expect_equal(c(terra::values(s1)), c(1, 0, 1, 0, NA))
   expect_equal(terra::values(s1), terra::values(s2))
@@ -237,19 +237,19 @@ test_that("start_solution", {
   s1 <-
     p %>%
     add_gurobi_solver(gap = 0, verbose = FALSE) %>%
-    solve_fixed_seed()
+    solve()
   s2 <-
     p %>%
     add_gurobi_solver(
       gap = 0, verbose = FALSE, start_solution = start_valid
     ) %>%
-    solve_fixed_seed()
+    solve()
   s3 <-
     p %>%
     add_gurobi_solver(
       gap = 0, verbose = FALSE, start_solution = start_invalid
     ) %>%
-    solve_fixed_seed()
+    solve()
   # test for correct solution
   expect_equal(c(terra::values(s1)), c(1, 0, 1, 0, NA))
   expect_equal(terra::values(s1), terra::values(s2))
@@ -280,9 +280,62 @@ test_that("node_file_start", {
       ) %>%
     add_gurobi_solver(gap = 0, verbose = FALSE, node_file_start = 0)
   # solve problem
-  s1 <- solve_fixed_seed(p)
-  s2 <- solve_fixed_seed(p)
+  s1 <- solve(p)
+  s2 <- solve(p)
   # test for correct solution
   expect_equal(c(terra::values(s1)), c(1, 0, 1, 0, NA))
   expect_equal(terra::values(s1), terra::values(s2))
+})
+
+test_that("solver information (single solution)", {
+  skip_on_cran()
+  skip_if_not_installed("gurobi")
+  # load data
+  sim_pu_raster <- get_sim_pu_raster()
+  sim_features <- get_sim_features()
+  # create problem
+  p <-
+    problem(sim_pu_raster, sim_features) %>%
+    add_min_set_objective() %>%
+    add_relative_targets(0.1) %>%
+    add_binary_decisions() %>%
+    add_gurobi_solver(time_limit = 5, verbose = FALSE)
+  # solve problem
+  s <- solve(p)
+  # tests
+  expect_true(is.numeric(attr(s, "objective")))
+  expect_length(attr(s, "objective"), 1)
+  expect_true(is.numeric(attr(s, "runtime")))
+  expect_length(attr(s, "runtime"), 1)
+  expect_true(is.character(attr(s, "status")))
+  expect_length(attr(s, "status"), 1)
+  expect_true(is.numeric(attr(s, "gap")))
+  expect_length(attr(s, "gap"), 1)
+})
+
+test_that("solver information (multiple solutions)", {
+  skip_on_cran()
+  skip_if_not_installed("gurobi")
+  # load data
+  sim_pu_raster <- get_sim_pu_raster()
+  sim_features <- get_sim_features()
+  # create problem
+  p <-
+    problem(sim_pu_raster, sim_features) %>%
+    add_min_set_objective() %>%
+    add_relative_targets(0.1) %>%
+    add_binary_decisions() %>%
+    add_shuffle_portfolio(3, remove_duplicates = FALSE) %>%
+    add_gurobi_solver(time_limit = 5, verbose = FALSE)
+  # solve problem
+  s <- solve(p)
+  # tests
+  expect_true(is.numeric(attr(s, "objective")))
+  expect_length(attr(s, "objective"), 3)
+  expect_true(is.numeric(attr(s, "runtime")))
+  expect_length(attr(s, "runtime"), 3)
+  expect_true(is.character(attr(s, "status")))
+  expect_length(attr(s, "status"), 3)
+  expect_true(is.numeric(attr(s, "gap")))
+  expect_length(attr(s, "gap"), 3)
 })